+++ playwright.config.ts
@@ -12,6 +12,14 @@ export default defineConfig({
     {
       name: 'Desktop Chromium',
       use: { ...devices['Desktop Chrome'] }
+    },
+    {
+      name: 'Desktop Firefox',
+      use: { ...devices['Desktop Firefox'] }
+    },
+    {
+      name: 'Desktop Safari',
+      use: { ...devices['Desktop Safari'] }
     }
   ],
   webServer: {